requirements.txt
PyQt6>=6.5.0
psutil>=5.9.0
pycryptodome>=3.18.0
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
cryptography>=41.0.0
requests>=2.28.0
pillow>=10.0.0
reportlab>=4.0.0
weasyprint>=59.0 